Moved StringUtils into matchbox package implicits.
