+++ src/test/scala/io/archivesunleashed/matchbox/StringUtilsTest.scala
@@ -14,11 +14,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.archivesunleashed.util
+package io.archivesunleashed.matchbox
 
 import java.io.IOException
 
-import io.archivesunleashed.util.StringUtils._
 import org.junit.runner.RunWith
 import org.scalatest.FunSuite
 import org.scalatest.junit.JUnitRunner
